requirements.txt
annotated-types==0.7.0
anyio==4.7.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
distlib==0.3.9
dnspython==2.7.0
exceptiongroup==1.2.2
fastapi==0.115.4
filelock==3.16.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
motor==3.6.0
packaging==24.2
pipenv==2024.4.0
platformdirs==4.3.6
pluggy==1.5.0
pydantic==2.10.4
pydantic_core==2.27.2
pymongo==4.9.2
pytest==8.3.3
python-dotenv==1.0.1
requests==2.32.3
setuptools==75.5.0
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
virtualenv==20.27.1
